requirements.txt
Django==2.0.7
django-import-export==1.0.1
django-model-utils==3.1.2
numpy==1.15.0
pandas==0.23.3
